Add botão de confirmação nas exclusões

+++ resources/views/alunos.blade.php
@@ -43,7 +43,7 @@
                                     <td>{{ $User->telefone }}</td>
                                     <td>
                                         <a class="button" href="/editar-aluno/{{ $User->id }}">Editar</a>
-                                        <a class="button" href="/excluir-aluno/{{ $User->id }}">Excluir</a>
+                                        <a class="button" onclick="return confirm('Tem certeza de que deseja excluir este aluno?')" href="/excluir-aluno/{{ $User->id }}">Excluir</a>
                                     </td>
                                 </tr>
                             @endforeach
